requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
pydantic==2.6.1
requests==2.31.0
python-dotenv==1.0.0
httpx==0.26.0

# Testing Dependencies
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
